fundacion/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
distlib==0.3.5
Django==4.1
django-ckeditor==6.5.0
django-js-asset==2.0.0
filelock==3.7.1
mysql-connector-python==8.0.29
mysqlclient==2.1.1
Pillow==9.2.0
platformdirs==2.5.2
PyMySQL==1.0.2
pyodbc==4.0.34
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.16.2
